[BE] refactor(#63): user 패키지 이동
* account 패키지 밑으로 이동
