+++ backend/src/main/java/com/example/backend/auth/api/service/oauth/adapter/google/GoogleAdapter.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import java.net.URI;
-import static com.example.backend.domain.define.user.constant.UserPlatformType.GOOGLE;
+import static com.example.backend.domain.define.account.user.constant.UserPlatformType.GOOGLE;
 
 @Slf4j
 @Component
