+++ backend/src/test/java/com/example/backend/auth/api/service/oauth/OAuthServiceTest.java
@@ -17,7 +17,6 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 import java.util.List;
 
-import static com.example.backend.domain.define.user.constant.UserPlatformType.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
